change nbelement fields to long ints

+++ lib/DB_File.h
@@ -10,6 +10,6 @@ typedef struct{
 }t_datablock;
 
 int Create_DB(dbc *db, char filename[]);
-long create_index_file(dbc* db, t_algo_meta* meta, int nb, t_datablock* i_block, t_datablock* t_block);
+long create_index_file(dbc* db, t_algo_meta* meta, long nb, t_datablock* i_block, t_datablock* t_block);
 
 #endif // DB_FILE_H_INCLUDED
